a bunch of documentation and code cleanup
